# e_learning_project-openapi.yaml
openapi: "3.1.0"
info:
  title: "e_learning_project API"
  description: "e_learning_project API"
  version: "1.0.0"
servers:
  - url: "https://e_learning_project"
paths:
...
  /users/count:
    get:
      summary: "GET users/count"
      operationId: "countUsersWithParams"
      parameters:
        - name: "lastName"
          in: "query"
          required: false
          schema:
            type: "string"
        - name: "firstName"
          in: "query"
          required: false
          schema:
            type: "string"
        - name: "email"
          in: "query"
          required: false
          schema:
            type: "string"
        - name: "username"
          in: "query"
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "integer"
                format: "int32"
  /users/count/search:
    get:
      summary: "GET users/count/search"
      operationId: "countUsersWithSearch"
      parameters:
        - name: "search"
          in: "query"
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "integer"
                format: "int32"
  /users/search:
    get:
      summary: "GET users/search"
      operationId: "searchUsersWithSearch"
      parameters:
        - name: "username"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "exact"
          in: "query"
          required: false
          schema:
            type: "boolean"
            default: "false"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/UserInfo"
...
  /upload:
    post:
      summary: "POST upload"
      operationId: "upload"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: "object"
              required:
                - "file"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ObjectUrl"
    delete:
      summary: "DELETE upload"
      operationId: "deleteAll"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjectUrls"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
...
  /published-courses/{courseId}:
    get:
      summary: "GET published-courses/{courseId}"
      operationId: "getPublishedCourseById"
      parameters:
        - name: "courseId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    put:
      summary: "PUT courses/{courseId}"
      operationId: "updateCourse"
      parameters:
        - name: "courseId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdateDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      summary: "DELETE courses/{courseId}"
      operationId: "deleteCourse"
      parameters:
        - name: "courseId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "force"
          in: "query"
          required: false
          schema:
            type: "boolean"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Void"
    get:
      summary: "GET courses/{courseId}"
      operationId: "getCourseById"
      parameters:
        - name: "courseId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Course"
...
  /discounts/{id}:
    put:
      summary: "PUT discounts/{id}"
      operationId: "updateDiscount"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Discount"
    delete:
      summary: "DELETE discounts/{id}"
      operationId: "deleteDiscount"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "force"
          in: "query"
          required: false
          schema:
            type: "boolean"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Void"
    get:
      summary: "GET discounts/{id}"
      operationId: "getDiscountById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Discount"
  /discounts:
    post:
      summary: "POST discounts"
      operationId: "createDiscount"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountDTO"
        required: true
      responses:
        "201":
          description: "Created"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Discount"
    get:
      summary: "GET discounts"
      operationId: "getAllDiscounts"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Discount"
  /discounts/{id}/restore:
    post:
      summary: "POST discounts/{id}/restore"
      operationId: "restoreDiscount"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Void"
  /discounts/code/{code}:
    get:
      summary: "GET discounts/code/{code}"
      operationId: "getDiscountByCode"
      parameters:
        - name: "code"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Discount"
...
components:
  schemas:
...
    MonetaryAmount:
      type: "object"
      properties: { }
...
    Instant:
      type: "object"
      properties: { }
...
    UserInfo:
      type: "object"
      properties:
        id:
          type: "string"
        firstName:
          type: "string"
        lastName:
          type: "string"
        username:
          type: "string"
        email:
          type: "string"
...
    ObjectUrl:
      type: "object"
      properties:
        url:
          type: "string"
    ObjectUrls:
      type: "object"
      properties:
        urls:
          type: "array"
          items:
            type: "string"
    Lesson:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        title:
          type: "string"
        type:
          type: "string"
          enum:
            - "VIDEO"
            - "TEXT"
            - "QUIZ"
            - "ASSIGNMENT"
        link:
          type: "string"
        quiz:
          type: "integer"
          format: "int64"
    CourseSection:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        title:
          type: "string"
        lessons:
          type: "array"
          items:
            $ref: "#/components/schemas/Lesson"
    StudentRef:
      type: "object"
      properties:
        student:
          type: "integer"
          format: "int64"
        firstName:
          type: "string"
        lastName:
          type: "string"
        email:
          type: "string"
    CourseRequest:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        type:
          type: "string"
          enum:
            - "PUBLISH"
            - "UNPUBLISH"
        status:
          type: "string"
          enum:
            - "PENDING"
            - "APPROVED"
            - "REJECTED"
        resolved:
          type: "boolean"
        resolvedBy:
          type: "string"
        requestedBy:
          type: "string"
        message:
          type: "string"
        rejectReason:
          type: "string"
        approveMessage:
          type: "string"
    Course:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        title:
          type: "string"
        thumbnailUrl:
          type: "string"
        description:
          type: "string"
        language:
          type: "string"
          enum:
            - "ENGLISH"
            - "VIETNAMESE"
            - "JAPANESE"
            - "KOREAN"
            - "CHINESE"
            - "FRENCH"
            - "GERMAN"
            - "SPANISH"
        subtitles:
          type: "array"
          items:
            type: "string"
            enum:
              - "ENGLISH"
              - "VIETNAMESE"
              - "JAPANESE"
              - "KOREAN"
              - "CHINESE"
              - "FRENCH"
              - "GERMAN"
              - "SPANISH"
        benefits:
          type: "array"
          items:
            type: "string"
        prerequisites:
          type: "array"
          items:
            type: "string"
        sections:
          type: "array"
          items:
            $ref: "#/components/schemas/CourseSection"
        price:
          $ref: "#/components/schemas/MonetaryAmount"
        published:
          type: "boolean"
        unpublished:
          type: "boolean"
        teacher:
          type: "string"
        approvedBy:
          type: "string"
        students:
          type: "array"
          items:
            $ref: "#/components/schemas/StudentRef"
        courseRequests:
          type: "array"
          items:
            $ref: "#/components/schemas/CourseRequest"
        version:
          type: "integer"
          format: "int32"
        createdBy:
          type: "string"
        createdDate:
          $ref: "#/components/schemas/Instant"
        lastModifiedBy:
          type: "string"
        lastModifiedDate:
          $ref: "#/components/schemas/Instant"
    CourseUpdateDTO:
      type: "object"
      properties:
        title:
          type: "string"
        description:
          type: "string"
        thumbnailUrl:
          type: "string"
        benefits:
          type: "array"
          items:
            type: "string"
        prerequisites:
          type: "array"
          items:
            type: "string"
        subtitles:
          type: "array"
          items:
            type: "string"
            enum:
              - "ENGLISH"
              - "VIETNAMESE"
              - "JAPANESE"
              - "KOREAN"
              - "CHINESE"
              - "FRENCH"
              - "GERMAN"
              - "SPANISH"
...
    Void:
      type: "object"
      properties: { }
...
    DiscountDTO:
      type: "object"
      properties:
        code:
          type: "string"
        type:
          type: "string"
          enum:
            - "PERCENTAGE"
            - "FIXED"
        percentage:
          type: "number"
          format: "double"
        fixedPrice:
          $ref: "#/components/schemas/MonetaryAmount"
        startDate:
          type: "string"
          format: "date-time"
        endDate:
          type: "string"
          format: "date-time"
        maxUsage:
          type: "integer"
          format: "int32"
    Discount:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        code:
          type: "string"
        type:
          type: "string"
          enum:
            - "PERCENTAGE"
            - "FIXED"
        percentage:
          type: "number"
          format: "double"
        fixedPrice:
          $ref: "#/components/schemas/MonetaryAmount"
        startDate:
          type: "string"
          format: "date-time"
        endDate:
          type: "string"
          format: "date-time"
        currentUsage:
          type: "integer"
          format: "int32"
        maxUsage:
          type: "integer"
          format: "int32"
        createdBy:
          type: "string"
        createdDate:
          $ref: "#/components/schemas/Instant"
        lastModifiedBy:
          type: "string"
        lastModifiedDate:
          $ref: "#/components/schemas/Instant"